Improve file structure and code readability
