+++ frontend/src/pages/admin/AdminBookings.css
